fix 6152

+++ far/filelist.cpp
@@ -5623,7 +5623,7 @@ bool FileList::PopPlugin(int EnableRestoreViewMode)
 		return false;
 	}
 
-	const std::wstring cached_hostfile = m_CachedOpenPanelInfo.HostFile;
+	const std::wstring cached_hostfile = m_CachedOpenPanelInfo.HostFile ? m_CachedOpenPanelInfo.HostFile : L"";
 	const auto cached_Flags = m_CachedOpenPanelInfo.Flags;
 	const auto CurPlugin = std::move(PluginsList.back());
 	PluginsList.pop_back();
@@ -5669,13 +5669,13 @@ bool FileList::PopPlugin(int EnableRestoreViewMode)
 
 
 		Global->CtrlObject->Plugins->GetOpenPanelInfo(GetPluginHandle(), &m_CachedOpenPanelInfo);
-		if (!(m_CachedOpenPanelInfo.Flags & OPIF_REALNAMES)) // remove previous plugin host-file/directory
+		if (!(m_CachedOpenPanelInfo.Flags & OPIF_REALNAMES) && !CurPlugin->m_HostFile.empty()) // remove previous plugin host-file/directory
 		{
 			char del_mode = 'd'; // old way - always remove directory on plugin panel pop
 			if ((cached_Flags & OPIF_RECURSIVEPANEL) && !(cached_Flags & OPIF_DELETEDIRONCLOSE)) // new way - controlled by flags
 				del_mode = (cached_Flags & OPIF_DELETEFILEONCLOSE) ? 'f' : '\0';
 
-			if (cached_hostfile == CurPlugin->m_HostFile)
+			if (cached_hostfile == CurPlugin->m_HostFile) // just in case
 			{
 				if (del_mode == 'd') DeleteFileWithFolder(CurPlugin->m_HostFile);
 				if (del_mode == 'f') std::ignore = os::fs::delete_file(CurPlugin->m_HostFile);
